pyproject.toml
[tool.poetry]
name = "langchain-genai"
version = "0.1.0"
description = ""
authors = ["muhammad-stack <142173430+muhammad-stack@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.8"
langchain-community = "^0.3.1"
langchain-google-genai = "^2.0.0"
python-dotenv = "^1.0.1"
langserve = {extras = ["all"], version = "^0.3.0"}
fastapi = "^0.115.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
